docs: remove note on reverse translations
This can cause schema violations in earlier versions. If a new
parameter or value is added to a given transport type in a later
version a decision on its handling may be taken at that point, but
it is suggested at present that any implementations which absolutely
require a parameter or value which is only exposed in a later version
do not translate these to earlier versions as it may be impossible.