Clarify source_ip behaviour
The `source_ip` parameter should always be set based on the `source-filter`, not the Origin (o=) line. The Origin unicast address may be a management IP address or obfuscated, according to RFC 4566, and moreover there is only one, so it makes a poor default in the case of ST 2022-7 for example. Examples are therefore adjusted:
1. To make the inclusive source-filter used to specify a source-specific multicast (SSM) session as per RFC 4570 different than the Origin unicast address.
2. In the case of any-source multicast, to set the `source_ip` to `null` as per the IS-05 APIs/schemas/receiver_transport_params_rtp.json description.
3. In the case of unicast operation, to set the `source_ip` to `null`. An inclusive source-filter would set the `source_ip`, but without one, I believe the `source_ip` should (or may?) be set to `null` to avoid the issues with the Origin unicast address mentioned already, despite the schema description.