Merge v1.1.x into v1.2-dev (#178)
* Fix copy/paste typo as per #169

* Remove duplicate keys from schema defintions (#176)

* Clarify what "requested_time" should be set to when "activation":"mode" is set to "activate_immediate" (#175)

It's unclear what "requested_time" should be set to when "activation":"mode" is set to "activate_immediate".  Should it be null or absent or either.  The examples show it set to null but absent doesn't seem unreasonable so I've gone with that option. Please amend as required.

* Behaviour: RTCP (#168)

* Update Behaviour - RTP Transport Type.md

Revised language in ARG call 2025-03-26

---------

Co-authored-by: Peter Brightwell <peter.brightwell@bbc.co.uk>

---------

Co-authored-by: Jonas Reucher <JonasReucher@gmail.com>
Co-authored-by: Pete-Cordell <pcordell@nevion.com>
Co-authored-by: N-Nagorny <15526762+N-Nagorny@users.noreply.github.com>